Revision of refactor_main branch
