+++ INCHI-1-SRC/INCHI_EXE/inchi-1/src/CMakeLists.txt
@@ -25,10 +25,10 @@ include_directories(${P_BASE} ${P_CURRENT})
 add_executable(inchi-1)
 
 target_sources(inchi-1 PRIVATE
-	main.c
-	${P_BASE}/ichimain.c
-	${P_BASE}/dispstru.c
-	${P_BASE}/dispstru.h
+	inchi_main_cli.c
+	ichimain.c
+	dispstru.c
+	dispstru.h
 	${P_BASE}/bcf_s.h
 	${P_BASE}/bcf_s.c
 	${P_BASE}/extr_ct.h
