+++ INCHI-1-SRC/INCHI_EXE/inchi-1/src/dispstru.h
