+++ INCHI-1-SRC/INCHI_EXE/inchi-1/vc14/inchi-1.sln
@@ -1,4 +1,4 @@
-
+﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25420.1
